maj line_edition
