+++ includes/sh.h
@@ -1,0 +1,6 @@
+#ifndef SH_H
+# define SH_H
+
+void	set_signal_ign(void);
+
+#endif
