+++ includes/struct.h
@@ -222,7 +222,7 @@ typedef struct	s_var
 
 typedef struct	s_pipe
 {
-	int32_t	fd[2];
+	int32_t		fd[2];
 	uint8_t		tmp;
 }				t_pipe;
 
@@ -232,8 +232,9 @@ typedef struct	s_process
 	char **av;                  /* for exec */
 	char *path;					/* path's exec */
 	pid_t pid;                  /* process ID */
-	uint8_t completed;             /* true if process has completed */
-	uint8_t stopped;               /* true if process has stopped */
+	uint8_t completed;          /* true if process has completed */
+	uint8_t stopped;            /* true if process has stopped */
+	uint8_t retour;				/* WEXITSTATUS  */
 	int status;                 /* reported status value */
 	uint8_t std[3];				/* stdin out err*/
 	uint8_t setup;
@@ -241,15 +242,15 @@ typedef struct	s_process
 
 typedef struct	s_job
 {
-	char *command;              /* command line, used for messages */
-	t_list *process;     		/* list of processes in this job */
-	t_list *var;				/* VAR env | locale | tmp */
-	pid_t pgid;                 /* process group ID */
-	uint8_t fg;					/* foreground */
-	t_pipe	pipe;
-	int  wstatus;				/* return of waitpid */
-	char notified;              /* true if user told about stopped job */
-	struct termios tmodes;      /* saved terminal modes */
+	char		*command;           /* command line, used for messages */
+	t_list		*process;     		/* list of processes in this job */
+	t_list		*var;				/* VAR env | locale | tmp */
+	pid_t		pgid;               /* process group ID */
+	uint8_t		fg;					/* foreground */
+	t_pipe		pipe;				/* pipeline */
+	char 		notified;           /* true if user told about stopped job */
+	uint8_t		std[3];				/* stdin out err*/
+	struct		termios tmodes;     /* saved terminal modes */
 } 				t_job;
 
 #endif
