+++ srcs/evaluator/launcher.c
@@ -15,12 +15,15 @@ void	ex(char *s)
 #include <signal.h>
 void	set_signal_child(void)
 {
+	signal(SIGHUP, SIG_DFL);
 	signal(SIGINT, SIG_DFL);
 	signal(SIGQUIT, SIG_DFL);
+	signal(SIGPIPE, SIG_DFL);
+	signal(SIGCHLD, SIG_DFL);
 	signal(SIGTSTP, SIG_DFL);
+	signal(SIGCONT, SIG_DFL);
 	signal(SIGTTIN, SIG_DFL);
 	signal(SIGTTOU, SIG_DFL);
-	signal(SIGCHLD, SIG_DFL);
 }
 
 /*    plus besoin si strjoin valide avec stdarg
@@ -79,32 +82,25 @@ int		builtin_process(void)
 	return (0);
 }
 
-int		wait_job(t_cfg *shell, t_job *job, pid_t wait)
-{
-	pid_t pid_child;
-
-	(void)shell;
-	pid_child = waitpid(wait, &job->wstatus, WUNTRACED);// wait le dernier dunjob
-	//pid_child = waitpid(-job->pgid, &job->wstatus, WUNTRACED);
-	printf("sort du wait\n");
-	return (0);
-}
 
 int		parent_process(t_job *job, t_process *process, int fd_pipe, char **envp)
 {
-	
+
 	if (fd_pipe)
 		if (close(fd_pipe) == -1)
 			ex("[Parent process] close error:");
 	if (job->pgid == 0)
 		job->pgid = process->pid;
-	
-	/*setpgid(process->pid, job->pgid);			//wait job controler
-	 *if (job->fg)
-	 *    tcsetpgrp(STDIN_FILENO, job->pgid);  */
+	process->status = RUNNING;
 
+	printf("\t[PARANT PROCESS]  PID  = %d\t PGID = %d \n", process->pid, job->pgid);
+	setpgid(process->pid, job->pgid);
+
+	if (job->fg)
+		if (tcsetpgrp(STDIN_FILENO, job->pgid))
+			perror("[PARENT PROCESS] error tcsetpgrp");
 	ft_del_tab((void **)envp);
-	return (process->pid);   //debug pour wait last processjob
+	return (0);
 }
 
 int		child_process(t_job *job, t_process *process, int fd_pipe, char **envp)
@@ -115,17 +111,18 @@ int		child_process(t_job *job, t_process *process, int fd_pipe, char **envp)
 
 	if (fd_pipe)
 		if (close(fd_pipe) == -1)
-			ex("[Parent process] close error:");
+			ex("[child process] close error:");
 	process->pid = getpid();
 	if (job->pgid == 0)
 		job->pgid = process->pid;
 	set_signal_child();
 	do_dup(process);
-	 /*setpgid(process->pid, job->pgid); 		//not do if !fg ??
-	 *if (job->fg)
-	 *    tcsetpgrp(STDIN_FILENO, job->pgid); */
+	setpgid(process->pid, job->pgid); 		//not do if !fg ??
+	 if (job->fg)
+		if (tcsetpgrp(STDIN_FILENO, job->pgid))
+			perror("[CHILD PROCESS] error tcsetpgrp");
 
-	printf("lanchement d un process\n [%s]", process->cmd);
+	printf("\t[CHILD PROCESS EXEC] \t[%s]\n\n\n", process->av[0]);
 	if ((execve(process->path, process->av, envp)) == -1)
 		ex("execve:");
 	return (0);
@@ -135,7 +132,7 @@ int		fork_process(t_job *job, t_process *process)
 {
 	char	**envp;
 
-
+	printf("\t[FORK PROCESS] [%s]\n", process->path);
 	envp = create_tab_env(job->var, 0); //problematique, a voir ac l'assignement
 	if ((process->pid = fork()) == -1)
 		ex("fork:");
@@ -151,17 +148,17 @@ int		run_process(t_cfg *shell, t_job *job, t_process *process)
 
 	//type_process ??
 	//faire les redir et open
+	printf("\t[RUN PROCESS]\n");
 	(void)shell;
 	int debug = process_type(job->var, process); //return (127 si pas trouver)
 	if (debug == 127) //pour eviter de lancer vide pour le moment
 		return (1);
-	printf("process_type retour = [%d]\n", debug);
+	//printf("process_type retour = [%d]\n", debug);
 	return (fork_process(job, process));
 }
 
 int		run_job(t_cfg *shell, t_job *job, t_list *process)
 {
-	/* Add : Loop until we are in the foreground.  */
 	//set le terminal et sauv ?
 	//https://www.gnu.org/software/libc/manual/html_node/Functions-for-Job-Control.html#Functions-for-Job-Control
 	//faire les pipes ?
@@ -171,27 +168,32 @@ int		run_job(t_cfg *shell, t_job *job, t_list *process)
 	//exec_type ?
 	//https://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_09_01_01
 
-	pid_t wait; //debug
+	printf("\t[RUN JOB]  PID 21SH = [%d]\n", shell->pid);
 
-		while (process)
+	while (process)
 	{
 		routine_set_pipe(process, &job->pipe);
-		wait = run_process(shell, job, process->data);
+		run_process(shell, job, process->data);
 		process = process->next;
 		if (job->pipe.tmp)
 			if (close(job->pipe.tmp) == -1)
 				ex("[check and do pipe] close error:");
 	}
 	if (job->fg)
-		wait_job(shell, job, wait);
+	{
+		wait_process(job);
+		if (tcsetpgrp(STDIN_FILENO, shell->pid))
+			perror("[run job] error tcsetpgrp");
+	}
+
 	/*
 	 *   if (!shell_is_interactive)
-    wait_for_job (j);
-  else if (foreground)
-    put_job_in_foreground (j, 0);
-  else
-    put_job_in_background (j, 0);
-											https://www.gnu.org/software/libc/manual/html_node/Stopped-and-Terminated-Jobs.html#Stopped-and-Terminated-Jobs
-	*/
+	 wait_for_job (j);
+	 else if (foreground)
+	 put_job_in_foreground (j, 0);
+	 else
+	 put_job_in_background (j, 0);
+https://www.gnu.org/software/libc/manual/html_node/Stopped-and-Terminated-Jobs.html#Stopped-and-Terminated-Jobs
+*/
 	return (0);
 }
