+++ srcs/init_shell.c
@@ -1,4 +1,6 @@
 #include "exec.h"
+#include "sh.h"
+#include "struct.h"
 #include "libft.h"
 #include <sys/types.h>
 #include <termios.h>
@@ -11,8 +13,10 @@ void		set_signal_ign(void)
 	signal(SIGQUIT, SIG_IGN);
 	signal(SIGTSTP, SIG_IGN);
 	signal(SIGTTIN, SIG_IGN);
+	//signal(SIGTTOU, SIG_DFL);
 	signal(SIGTTOU, SIG_IGN);
-	signal(SIGCHLD, SIG_IGN);
+	signal(SIGCHLD, SIG_DFL);  //only for exec ?
+	//signal(SIGCHLD, SIG_IGN);
 }
 
 void		init_cfg(t_cfg *cfg, char **env)
@@ -35,8 +39,9 @@ void		init_shell(t_cfg *cfg, char **env)
 			kill(-shell_pgid, SIGTTIN);
 		set_signal_ign();
 		if (setpgid(cfg->pid, cfg->pid) < 0)
-			ex("setpgid");
-		tcsetpgrp(cfg->pid, cfg->pid);
+			ex("[INIT SHELL] error setpgid");
+		if (tcsetpgrp(STDIN_FILENO, cfg->pid))
+			perror("[INIT SHELL] error tcsetpgrp");
 		tcgetattr(STDIN_FILENO, &cfg->term_origin);
 	}
