+++ srcs/line_edition/term_init.c
@@ -6,6 +6,7 @@
 /*   By: ambelghi <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/13 14:23:13 by ambelghi          #+#    #+#             */
+/*   Updated: 2020/02/13 22:08:56 by hmerieux         ###   ########.fr       */
 /*   Updated: 2020/02/13 15:53:35 by ambelghi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
